save only relevant data items while parsing PDBx dictionary
