+++ converter.go
@@ -1,279 +1,27 @@
 package main
 
 import (
-	converter "converter/converterUtils"
-	"encoding/csv"
-	"encoding/json"
+	"converter/converterUtils"
+	"converter/parser"
 	"fmt"
 	"log"
 	"os"
 	"strings"
 )
 
-func getKeys[K string, V any](m map[string]V) []string {
-	keys := make([]string, 0)
-	for k := range m {
-		keys = append(keys, k)
-	}
-	return keys
-}
-
-func keyConcat(s1, s2 string) string {
-	if s1 != "" {
-		return s1 + "." + s2
-	} else {
-		return s2
-	}
-}
-
-func itemParent(item string) string {
-	return strings.Split(item, ".")[0]
-}
-
-func findElemByItem(item string, mapper map[string]string, toMmCIF bool) []string {
-	elements := make([]string, len(mapper))
-	if toMmCIF {
-		i := 0
-		for e := range mapper {
-			elements[i] = mapper[e]
-			i++
-		}
-	} else {
-		elements = getKeys(mapper)
-	}
-	var simElem []string
-	for i := range elements {
-		if itemParent(elements[i]) == item {
-			simElem = append(simElem, elements[i])
-		}
-	}
-	return simElem
-}
-func visit(mymap map[string]any, my_json_path string, my_long_properties map[string]any, lenVal int, valI int) map[string]any {
-	keys := getKeys(mymap)
-	for _, k := range keys {
-		jsonFlat := keyConcat(my_json_path, k)
-		switch nestedMap := mymap[k].(type) {
-		case map[string]any:
-			visit(nestedMap, jsonFlat, my_long_properties, lenVal, valI)
-		case []any:
-			for i, jsonSlice := range nestedMap {
-				if js, ok := jsonSlice.(map[string]any); ok {
-					visit(js, jsonFlat, my_long_properties, len(nestedMap), i)
-				} else {
-					log.Fatal("Weird case ")
-				}
-			}
-		default:
-			if lenVal == 1 {
-				my_long_properties[jsonFlat] = fmt.Sprint(mymap[k])
-			} else {
-				if valI == 0 {
-					values := make([]string, lenVal)
-					values[0] = fmt.Sprint(mymap[k])
-					my_long_properties[jsonFlat] = values
-				} else {
-					values := my_long_properties[jsonFlat]
-					if v, ok := values.([]string); ok {
-						v[valI] = fmt.Sprint(mymap[k])
-						my_long_properties[jsonFlat] = v
-					}
-				}
-			}
-		}
-	}
-	return my_long_properties
-}
-
-func readJson(jsonPath string) map[string]any {
-	// read json file and keep it in form of a map with keys consisting.of.joined.elements
-	var data []byte
-	data, err := os.ReadFile(jsonPath)
-	if err != nil {
-		log.Fatal("Error while reading the Json ", err)
-	}
-	// unmarshal json
-	var jsonContent map[string]any
-	json.Unmarshal([]byte(data), &jsonContent)
-
-	jsonFlat := make(map[string]any)
-	jsonFlat = visit(jsonContent, "", jsonFlat, 1, 0)
-
-	return jsonFlat
-}
-
-func formatMapper(path string, toMmCIF bool) map[string]string {
-	// read mapper file and based on the task (JSON -> mmCIF or mmCIF -> JSON) create a map with variable names
-	file, err := os.Open(path)
-	if err != nil {
-		log.Fatal("Error while reading the file ", err)
-	}
-	defer file.Close()
-	reader := csv.NewReader(file)
-	reader.Comma = '\t'
-	records, err := reader.ReadAll()
-	if err != nil {
-		log.Fatal("Error reading records", err)
-	}
-	// get the length of json map
-	l := 0
-	for i := range len(records) {
-		if records[i][1] != "" {
-			l++
-		}
-	}
-	mapper := make(map[string]string, 0)
-	if toMmCIF {
-		// json properties are mapped to the mmcif properties, missing ones are skipped
-		for _, eachrecord := range records {
-			if eachrecord[1] != "" {
-				mapper[eachrecord[0]] = eachrecord[1]
-			}
-		}
-	} else {
-		for _, eachrecord := range records {
-			if eachrecord[1] != "" {
-				mapper[eachrecord[1]] = eachrecord[0]
-			}
-		}
-	}
-	return mapper
-}
-func getKeyByValue(value string, m map[string]string) string {
-	for k, v := range m {
-		if v == value {
-			return k
-		}
-	}
-	return ""
-}
-
-func contains(s []string, e string) bool {
-	for _, a := range s {
-		if a == e {
-			return true
-		}
-	}
-	return false
-}
-
-func getLongest(s []string) int {
-	var r int
-	for _, a := range s {
-		if len(a) > r {
-			r = len(a)
-		}
-	}
-	return r
-}
-
-func valueMapper(nameMapper map[string]string, PDBxItems map[string][]string, valuesMap map[string]any) string {
-	// values from JSON are mapped to the mmcif properties
-	mappedVal := make([]string, 0, len(nameMapper))
-	var str strings.Builder
-	str.WriteString("data_myID\n#\n")
-	for jsonName := range valuesMap {
-		PDBxName := nameMapper[jsonName]
-		if PDBxName == "" {
-			continue // because translation is iterating on json, it still contains elements that don't exist in mmcif
-		}
-		var orderedElements = PDBxItems[itemParent(PDBxName)[1:]]
-		if contains(mappedVal, PDBxName) {
-			continue
-		}
-		elements := findElemByItem(itemParent(PDBxName), nameMapper, true)
-		var valueString string
-		switch valSlice := valuesMap[jsonName].(type) {
-		case []string: // loop notation
-			if valSlice == nil {
-				continue // not required in mmCIF
-			}
-			str.WriteString("loop_\n")
-
-			// list category names
-			// instead of all find elements use the ordered list from parsing the dictionary
-
-			for _, oE := range orderedElements {
-				for _, e := range elements {
-					if oE == strings.Split(e, ".")[1] {
-						fmt.Fprintf(&str, "%s\n", e)
-						mappedVal = append(mappedVal, e)
-					}
-				}
-			}
-			// write the values
-			for i := range len(valSlice) {
-				for _, oE := range orderedElements {
-					for _, e := range elements {
-						if oE == strings.Split(e, ".")[1] {
-							jsonKey := getKeyByValue(e, nameMapper)
-
-							if slice, ok := valuesMap[jsonKey].([]string); ok {
-								if strings.Contains(slice[i], " ") {
-									valueString = fmt.Sprintf("'%s' ", slice[i]) // if name contains whitespaces enclose it in single quotes
-								} else {
-									valueString = fmt.Sprintf("%s ", slice[i]) // take value as is
-								}
-							} else { // if name is present in both OSCEM and PDBx but no value is available set it as "omitted"
-								valueString = ". "
-							}
-							fmt.Fprintf(&str, "%s", valueString)
-						}
-					}
-				}
-				str.WriteString("\n")
-			}
-			str.WriteString("#\n")
-		case string: // simple list of categories
-			l := getLongest(elements) + 5
-			for _, oE := range orderedElements {
-				for _, e := range elements {
-					if oE == strings.Split(e, ".")[1] {
-
-						jsonKey := getKeyByValue(e, nameMapper)
-						if valuesMap[jsonKey] == nil {
-							continue // not required in mmCIF
-						}
-						formatString := fmt.Sprintf("%%-%ds", l)
-						fmt.Fprintf(&str, formatString, e)
-						if jsonValue, ok := valuesMap[jsonKey].(string); ok {
-
-							if strings.Contains(jsonValue, " ") {
-								valueString = fmt.Sprintf("'%s'\n", jsonValue)
-							} else {
-								valueString = fmt.Sprintf("%s\n", jsonValue)
-							}
-							fmt.Fprintf(&str, "%s", valueString)
-						}
-
-						mappedVal = append(mappedVal, e)
-					}
-				}
-			}
-			str.WriteString("#\n")
-		default:
-			fmt.Println("Problem appeared while unmarshalling JSON")
-		}
-	}
-	return str.String()
-}
-
 func main() {
 	jsonInstr := os.Args[1]
 	jsonSample := os.Args[2]
-	jsonToMmCif := os.Args[3]
-	dictFile, err := os.Open(os.Args[4])
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer dictFile.Close()
+
+	conversionFile := os.Args[3]
+	dictFile := os.Args[4]
 
 	mmCIFpath := os.Args[5]
 	unitsPath := os.Args[6]
 
 	// read all input json files and create one map from them all
-	mapInstr := readJson(jsonInstr)
-	mapSample := readJson(jsonSample)
+	mapInstr := parser.FromJson(jsonInstr)
+	mapSample := parser.FromJson(jsonSample)
 
 	// create one mapping for all the JSON contents: key - value
 	mapJson := make(map[string]any, len(mapInstr)+len(mapSample))
@@ -285,10 +33,10 @@ func main() {
 	}
 
 	// create a map containing OSCEM - PDBx naming mappings
-	mapper := formatMapper(jsonToMmCif, true)
+	mapper, OSCEMunits := parser.ConversionTable(conversionFile)
 
 	// parse PDBx dictionary to retrieve order of data items and units
-	dataItems, units := converter.ParseDict(dictFile)
+	dataItems, _ := parser.PDBxDict(dictFile, converterUtils.GetValues(mapper))
 
 	// use only a map of dataItems that will be needed my mapper
 	var PDBxdataItems = make(map[string][]string)
@@ -302,7 +50,7 @@ func main() {
 	}
 
 	// create mmCIF text
-	mmCIFlines := valueMapper(mapper, PDBxdataItems, mapJson)
+	mmCIFlines := parser.ToMmCIF(mapper, PDBxdataItems, mapJson)
 
 	// now write to cif file
 
@@ -334,7 +82,7 @@ func main() {
 	//var unitsString = ""
 	var unitsString strings.Builder
 	unitsString.WriteString("")
-	for k, v := range units {
+	for k, v := range OSCEMunits {
 		fmt.Fprintf(&unitsString, "%s,%s\n", k, v)
 	}
 	// Write the string to the file
@@ -345,4 +93,4 @@ func main() {
 	}
 }
 
-// go run converter.go ../OSCEM_Schemas/Instrument/test_data_valid.json ../OSCEM_Schemas/Sample/Sample_valid.json /Users/sofya/Documents/openem/converter-JSON-to-mmCIF/data/mapper.tsv ./data/mmcif_pdbx_v50.dic /Users/sofya/Documents/openem/converter-JSON-to-mmCIF/results/output.cif /Users/sofya/Documents/openem/converter-JSON-to-mmCIF/results/units.csv
+// go run converter.go data/data_instrument.json data/data_sample.json /Users/sofya/Documents/openem/LS_Metadata_reader/conversion/conversions.csv ./data/mmcif_pdbx_v50.dic /Users/sofya/Documents/openem/converter-JSON-to-mmCIF/results/output.cif /Users/sofya/Documents/openem/converter-JSON-to-mmCIF/results/units.csv
