+++ data/data_instrument.json
@@ -1,0 +1,55 @@
+{
+    "Instrument": {
+        "Microscope": "Titan",
+        "Illumination": "FloodBeam",
+        "Imaging": "Brightfield",
+        "Electron_source": "FEG",
+        "Acceleration_Voltage": 300,
+        "C2_Aperture": 70,
+        "CS": 2.7
+    },
+    "Acquisition":{
+        "Nominal_defocus": {
+            "Minimal": -1.0,
+            "Maximal": -2.5
+        },
+        "Calibrated_defocus": {
+            "Minimal": -1.0,
+            "Maximal": -2.5
+        },
+        "Nominal_magnification": 125000,
+        "Calibrated_magnification": 125666,
+        "Holder": "Autogrid holder",
+        "Holder_cryogen": "NaN",
+        "Temperature": {
+            "Minimal": -190,
+            "Maximal": -170
+        },
+        "Software": "SerialEM",
+        "Detector": "Falcon 4i",
+        "Detector_mode": "counting",
+        "Dose_per_movie": 50,
+        "Energy_filter": {
+            "Used": true,
+            "Model": "Selectris",
+            "Width": 10
+        },
+        "Image_size": {
+            "Height": 4096,
+            "Width": 4096
+        },
+        "Datetime": "today:4-5-2024",
+        "Exposure_time": 3,
+        "Tilt_angle": {
+            "Minimal": 0,
+            "Maximal": 0,
+            "Increment": 0
+        },
+        "Cryogen": "LN2",
+        "Frames_per_movie": 45,
+        "Grids_imaged": 1,
+        "Images_generated": 30000,
+        "Binning": 1,
+        "Pixel_size": 0.3
+    }
+}
