+++ data/data_sample.json
@@ -1,0 +1,86 @@
+{
+    "Grant": [
+        {
+            "Grant_ID": "SNF321",
+            "Funding_agency": "SNF"
+        },
+        {
+            "Grant_ID": "Fundingofsomekind",
+            "Funding_agency": "SNF"
+        }
+    ],
+    "Author": [
+        {
+            "FirstName": "John",
+            "LastName": "Doe",
+            "Email": "john.doe@gmail.com",
+            "Phone": "+4132112312",
+            "ORCID": "ORCID_123124151231",
+            "Organization": {
+                "Name_org": "University of Blub",
+                "Type_org": "Academic",
+                "Country": "Switzerland"
+            }
+            
+        },
+        {
+            "FirstName": "Jane",
+            "LastName": "Doe",
+            "Email": "jane.doe@gmail.com",
+            "Phone": "+4132112312",
+            "ORCID": "ORCID_123124151221",
+            "Organization": {
+                "Name_org": "University of Blub",
+                "Type_org": "Academic",
+                "Country": "Switzerland"
+            }
+        }
+    ],
+    "Sample": {
+        "Overall_Molecule": {
+            "Type": "Complex",
+            "Name": "Ribosome",
+            "Source": "yeast",
+            "Molecular_weight": 3000000
+        },
+        "Molecules": [
+            {
+                "Name": "Ribosome_",
+                "Type": "polypeptide",
+                "Class": "None of these",
+                "Sequence": "MMKNMMKYMMMYSYKEKSEKRRWMMKREMMKYLELLNMRMNKWVKRMMMSRKRLLSRMNRLEKKNHMNDYKLMSYNFNKNSMMMKSMLSKLMMNLLENVMNGVDLKSGKSKMSGGNMMMSKPVMKENLNTVNMLFYYFLPNNKSYKYFNRMNMYLNKHMKNYKKLVKLNKNYSKSSYLLQDKLMMSNMFNNMMFNRNNMKYMNNNVNNLMSSLNMNSNNKSLYLSLLNKTLNNNLSNLYSNNNINKSIINNTLVNMSMLNNNYNNNNTTFNINNNLFNLLNLLNNNYNNNNNYNNMMSNTNMLLNNYNNNNNNNNNNKCNMHSKMKLERKNMMLNYLLKNEMMNKNQMWMSKMENNKMNNINNNNNMLKMNKENDKTNMFGYMMSYMDMLLGNMMNKSKKSNDMKMMMSKYFGLKEVNMTGMNLKYEFNNTEMLLKLMRKGMSKRKRTLSRMFRFRLKNRMPLLNDKGMLKNKMSNNLLKNLALNNNILNVEYNIKNVLFNSINNNNINNNNNNNMYNDMKDYYNNMSLENKKDLSLNNELLYKNMVGWSLLLKGKVGARKGKNRSNRMLMTKGSFKNNNLYIYNIFDDNSNNNGYTKDRLRLNYMKNSNFISYMDKSTNNGKLGMTLKVNIL",
+                "Natural_source": "yeast",
+                "Taxonomy_ID_source": "GCA_003054445.1",
+                "Expression_system": "Pichia kudriavzevii",
+                "Taxonomy_ID_expression": "GCA_003054445.1",
+                "Gene_name": "Rib"
+            }
+        ],
+        "Ligands": [
+            {
+                "present": false
+            }
+        ],
+        "Specimen": {
+            "State": "frozen",
+            "Concentration": 0.3,
+            "pH": 7,
+            "Vitrification": false,
+            "Vitrification_cryogen": "None",
+            "Humidity": 90,
+            "Temperature": 300,
+            "Staining": false,
+            "Embedding": false,
+            "Shadowing": false
+        },
+        "Grid": {
+            "Manufacturer": "Quantifoil",
+            "Material": "Au",
+            "Mesh": 300,
+            "Film_support": false,
+            "Pretreatment_type": "Glow discharge",
+            "Pretreatment_time": 30,
+            "Pretreatment_pressure": 0.000001
+        }
+    }
+}
