--- parseDict/go.mod
@@ -1,3 +1,0 @@
-module dictionaryParser.go
-
-go 1.22.1
