+++ results/bla.txt
@@ -1,0 +1,2 @@
+map[Acquisition.Alignment_procedure:_em_imaging.alignment_procedure  Acquisition.Calibrated_defocus.Maximal:_em_imaging.calibrated_defocus_max Acquisition.Calibrated_defocus.Minimal:_em_imaging.calibrated_defocus_min Acquisition.Calibrated_magnification:_em_imaging.calibrated_magnification Acquisition.Cryogen:_em_imaging.cryogen Acquisition.Datetime:_em_imaging.date Acquisition.Detector:_em_image_recording.film_or_detector_model Acquisition.Detector_mode:_em_image_recording.detector_mode   Acquisition.Dose_per_movie:_em_image_recording.avg_electron_dose_per_image Acquisition.Energy_filter.Model:_em_imaging_optics.energyfilter_name Acquisition.Energy_filter.Width:_em_imaging_optics.energyfilter_slit_width Acquisition.Exposure_time:_em_image_recording.average_exposure_time Acquisition.Frames_per_movie:_em_image_scans.frames_per_image Acquisition.Grids_imaged:_em_image_recording.num_grids_imaged Acquisition.Holder:_em_imaging.specimen_holder_model Acquisition.Images_generated:_em_image_recording.num_real_images Acquisition.Microscope_software:_em_software.category ['IMAGE ACQUISITION'] Acquisition.Nominal_defocus.Maximal:_em_imaging.nominal_defocus_max Acquisition.Nominal_defocus.Minimal:_em_imaging.nominal_defocus_min Acquisition.Nominal_magnification:_em_imaging.nominal_magnification  Acquisition.Specialist_Optics.Chromatic_Aberration_Corrector.Used:_em_imaging_optics.chr_aberration_corrector Acquisition.Specialist_Optics.Phaseplate.Used:_em_imaging_optics.phase_plate Acquisition.Specialist_Optics.Spherical_Aberration_Corrector.Used:_em_imaging_optics.sph_aberration_corrector Acquisition.Temperature.Maximal:_em_imaging.recording_temperature_maximum Acquisition.Temperature.Minimal:_em_imaging.recording_temperature_minimum Acquisition.Tilt_angle.Increment:_em_imaging.tilt_angle_increment Acquisition.Tilt_angle.Maximal:_em_imaging.tilt_angle_max Acquisition.Tilt_angle.Minimal:_em_imaging.tilt_angle_min  Author.Email:_pdbx_contact_author.email Author.FirstName:_pdbx_contact_author.name_first Author.LastName:_pdbx_contact_author.name_last Author.ORCID:_pdbx_contact_author.identifier_ORCID Author.Organization.Country:_pdbx_contact_author.country Author.Organization.Type_org:_pdbx_contact_author.organization_type Author.Phone:_pdbx_contact_author.phone Author.Role:_pdbx_contact_author.role Grant.Funding_agency:_pdbx_audit_support.funding_organization Grant.Grant_ID:_pdbx_audit_support.grant_number  Instrument.Acceleration_Voltage:_em_imaging.accelerating_voltage  Instrument.C2_Aperture:_em_imaging.c2_aperture_diameter Instrument.CS:_em_imaging.nominal_cs  Instrument.Electron_source:_em_imaging.electron_source  Instrument.Illumination:_em_imaging.illumination_mode  Instrument.Imaging:_em_imaging.mode Instrument.Microscope:_em_imaging.microscope_model Sample.Grid.Film_material:_em_support_film.material Sample.Grid.Film_thickness:_em_support_film.thickness  Sample.Grid.Film_topology:_em_support_film.topology  Sample.Grid.Manufacturer:_em_sample_support.grid_type Sample.Grid.Material:_em_sample_support.grid_material Sample.Grid.Mesh:_em_sample_support.grid_mesh_size Sample.Grid.Pretreatment_atmosphere:_em_grid_pretreatment.atmosphere  Sample.Grid.Pretreatment_pressure:_em_grid_pretreatment.pressure Sample.Grid.Pretreatment_time:_em_grid_pretreatment.time Sample.Grid.Pretreatment_type:_em_grid_pretreatment.type Sample.Ligands.SMILE:_pdbx_chem_comp_depositor_info.descriptor Sample.Ligands.present:_pdbx_entry_details.has_ligand_of_interest Sample.Molecules.Class:_em_entity_assembly.type Sample.Molecules.Expression_system:_em_entity_assembly_recombinant.organism Sample.Molecules.Gene_name:_entity_src_gen.pdbx_gene_src_gene Sample.Molecules.Name:_pdbx_entity_branch_descriptor.descriptor Sample.Molecules.Natural_source:_em_entity_assembly_naturalsource.organism Sample.Molecules.Sequence:_entity_poly.pdbx_seq_one_letter_code Sample.Molecules.Taxonomy_ID_expression:_em_entity_assembly_recombinant.ncbi_tax_id Sample.Molecules.Taxonomy_ID_source:_em_entity_assembly_naturalsource.ncbi_tax_id  Sample.Molecules.Type:_entity_poly.type Sample.Overall_Molecule.Molecular_weight:_entity.formula_weight Sample.Overall_Molecule.Name:_em_entity_assembly.name Sample.Overall_Molecule.Source:_entity.src_method Sample.Overall_Molecule.Type:_entity.type  Sample.Specimen.Buffer:_em_buffer.details Sample.Specimen.Concentration:_em_specimen.concentration Sample.Specimen.Embedding:_em_specimen.embedding_applied  Sample.Specimen.Humidity:_em_vitrification.humidity  Sample.Specimen.Shadowing:_em_specimen.shadowing_applied Sample.Specimen.Staining:_em_specimen.staining_applied Sample.Specimen.Temperature:_em_vitrification.chamber_temperature Sample.Specimen.Vitrification:_em_specimen.vitrification_applied Sample.Specimen.Vitrification_cryogen:_em_vitrification.cryogen_name Sample.Specimen.pH:_em_buffer.pH]
+String successfully written to the file.
